@@ -18,7 +18,6 @@
     "react-native-svg": "^12.1.1",
     "react-native-web": "^0.17.1",
     "react-scripts": "5.0.1",
-    "typescript": "^4.8.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -47,6 +46,15 @@
     ]
   },
   "devDependencies": {
-    "react-router-dom": "^6.4.5"
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
+    "eslint": "^8.31.0",
+    "eslint-config-standard-with-typescript": "^27.0.1",
+    "eslint-plugin-import": "^2.27.4",
+    "eslint-plugin-n": "^15.6.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.32.0",
+    "react-router-dom": "^6.4.5",
+    "typescript": "^4.9.4"
   }
 }
