@@ -1,15 +1,16 @@
-import { HStack, VStack,  Box , Button, Text, Menu, Pressable, HamburgerIcon, Divider} from "native-base";
-import { useNavigate } from "react-router-dom";
-import theme from '../components/NativeBaseStyling'
-import { auth } from "../xplat/Firebase";
+import { HStack, VStack,  Box , Button, Text, Menu, Pressable, HamburgerIcon, Divider} from 'native-base';
+import { useNavigate } from 'react-router-dom';
+import theme from '../components/NativeBaseStyling';
+import { auth } from '../xplat/Firebase';
 
 export const NavBar = () => {
-    let navigate = useNavigate();
+    const navigate = useNavigate();
     return (
         <Box height='1' zIndex={100}>
             <VStack space={2} flex={1} backgroundColor={'#F1F1F1'} zIndex={100}>
-                <HStack space={1} alignItems='left' p={1} marginBottom={'5px'} width = '100%' backgroundColor={'#F1F1F1'} paddingTop={'5px'} position={'fixed'} zIndex={100}>
-                    <Button onPress={()=>{navigate('/routes')}}>
+                <HStack space={1} alignItems='left' p={1} marginBottom={'5px'} width = '100%' 
+                    backgroundColor={'#F1F1F1'} paddingTop={'5px'} position={'fixed'} zIndex={100}>
+                    <Button onPress={()=>{navigate('/routes');}}>
                         <Text variant={'button'}>
                             Routes
                         </Text>
@@ -32,13 +33,13 @@ export const NavBar = () => {
                     <Button onPress={() => navigate('/competitions')}>
                         <Text variant={'button'}>
                                 Competitions
-                            </Text>
+                        </Text>
                     </Button>
                     <Box top={'15px'} right='1%' position={'fixed'}>
                         <Menu  trigger={triggerProps => {
                             return <Pressable accessibilityLabel="Online session options" {...triggerProps} >
-                                    <HamburgerIcon size='lg' />
-                                    </Pressable>;
+                                <HamburgerIcon size='lg' />
+                            </Pressable>;
                         }}>
                             <Menu.Item onPress={() => {
                                 auth.signOut();
@@ -51,4 +52,4 @@ export const NavBar = () => {
         </Box>
         
     );
-}
+};
