@@ -5,16 +5,16 @@ import { Page } from '../App';
 const LandingPage = ({ setCurrentPage }: { setCurrentPage: (arg0: Page) => void; }) => {
 
     const backendOnLocalhost = () => {
-        if (window.location.hostname === "localhost") setCurrentPage(Page.BackendTesting)
-        else console.log("Why did you even know to click this?")
-    }
+        if (window.location.hostname === 'localhost') setCurrentPage(Page.BackendTesting);
+        else console.log('Why did you even know to click this?');
+    };
 
     return (
         <div className="App">
             <header className="App-header">
                 <img src={logo} className="App-logo" alt="logo" />
                 <p onClick={backendOnLocalhost}>
-                    Welcome to Tower! We, uh, don't have much for you to do here yet...
+                    Welcome to Tower! We, uh, dont have much for you to do here yet...
                 </p>
                 <a
                     className="App-link"
@@ -26,6 +26,6 @@ const LandingPage = ({ setCurrentPage }: { setCurrentPage: (arg0: Page) => void;
             </header>
         </div>
     );
-}
+};
 
 export default LandingPage;
