@@ -1,17 +1,17 @@
-import { NavBar } from "../components/NavigationBar";
-import { Box, Button, Text} from 'native-base'
-import { useNavigate, createSearchParams } from "react-router-dom";
+import { NavBar } from '../components/NavigationBar';
+import { Box, Button, Text} from 'native-base';
+import { useNavigate, createSearchParams } from 'react-router-dom';
 
 
 const Routes = () => {
-    let navigate = useNavigate();
+    const navigate = useNavigate();
     const exampleSearchParams = {uid: 'GQBdclAMmE2v4nDPphsc'};
     const navToRoute = () => {
         navigate({
             pathname: '/route',
             search: `?${createSearchParams(exampleSearchParams)}`
-        })
-    }
+        });
+    };
 
     return (
         <Box flexDir={'column'}>
@@ -25,6 +25,6 @@ const Routes = () => {
         
 
     );
-}
+};
 
 export default Routes;
