@@ -1,12 +1,12 @@
-import { Text, Box, Input, Button, HStack, FormControl, Modal} from "native-base";
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { createUser, isKnightsEmail, sendAuthEmail} from "../xplat/api";
-import { auth } from "../xplat/Firebase";
-import logo from '../logo.svg'
+import { Text, Box, Input, Button, HStack, FormControl, Modal} from 'native-base';
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { createUser, isKnightsEmail, sendAuthEmail} from '../xplat/api';
+import { auth } from '../xplat/Firebase';
+import logo from '../logo.svg';
 
 const Signup =  () => {
-    let navigate = useNavigate();
+    const navigate = useNavigate();
     const [displayName, setDisplayName] = useState('');
     const [username, setUsername] = useState('');
     const [email, setEmail] = useState('');
@@ -23,13 +23,13 @@ const Signup =  () => {
         setEmail('');
         setPassword('');
         setPasswordCheck('');
-    }
+    };
 
     const closeModal = () => {
         auth.signOut().then(() => {
             setShowModal(false);
-        })
-    }
+        });
+    };
 
     async function attemptSignup() {
         setSignupError(false);
@@ -52,8 +52,8 @@ const Signup =  () => {
             }
         }).catch( (error) => {
             setSignupError(true);
-        })
-    };
+        });
+    }
 
     return (
         <Box flex={1} alignSelf='center' alignItems={'center'} justifyContent='center' paddingTop= '10vh'>
@@ -64,15 +64,15 @@ const Signup =  () => {
                     <Modal.Header>Email Confirmation</Modal.Header>
                     <Modal.Body>
                         A confirmation email has been sent to verify your email address.{'\n'}
-                        If you are a Tower employee you will need to verify your email and then contact your manager to gain the appropriate
-                        permissions to access the website.
+                        If you are a Tower employee you will need to verify your email and then contact your manager 
+                        to gain the appropriate permissions to access the website.
                     </Modal.Body>
                     <Modal.Footer width={'fill'}>
                         <Button flex="1" onPress={() => {
                             setShowModal(false);
                             navigate('/');
                         }}>
-                        <Text variant={'button'}>Close</Text>
+                            <Text variant={'button'}>Close</Text>
                         </Button>
                     </Modal.Footer>
                 </Modal.Content>
@@ -83,19 +83,27 @@ const Signup =  () => {
             </Text>
             <FormControl alignItems={'center'} isInvalid={passwordError}>
                 <FormControl.Label isRequired>Email</FormControl.Label>
-                <Input isRequired onChangeText={(e) => setEmail(e)} marginBottom='5px' placeholder="email" width={'60%'}/>
+                <Input isRequired 
+                    onChangeText={(e) => setEmail(e)} marginBottom='5px' placeholder="email" width={'60%'}/>
                 <FormControl.HelperText>Must end with @knights.ucf.edu or @ucf.edu</FormControl.HelperText>
                 <FormControl.Label isRequired>Username</FormControl.Label>
-                <Input isRequired onChangeText={(e) => setUsername(e)} marginBottom='5px' placeholder="@username" width={'60%'}/>
+                <Input isRequired 
+                    onChangeText={(e) => setUsername(e)} marginBottom='5px' placeholder="@username" width={'60%'}/>
                 <FormControl.Label isRequired>Display Name</FormControl.Label>
-                <Input isRequired onChangeText={(e) => setDisplayName(e)} marginBottom='5px' placeholder="display name" width={'60%'}/>
+                <Input isRequired onChangeText={(e) => setDisplayName(e)} 
+                    marginBottom='5px' placeholder="display name" width={'60%'}/>
                 <FormControl.Label isRequired> Password</FormControl.Label>
-                <Input isRequired onChangeText={(e) => setPassword(e)} marginBottom='5px' type="password" placeholder="password" width={'60%'}/>
+                <Input isRequired onChangeText={(e) => setPassword(e)} 
+                    marginBottom='5px' type="password" placeholder="password" width={'60%'}/>
                 <FormControl.Label isRequired> Password Confirmation</FormControl.Label>
-                <Input isRequired onChangeText={(e) => setPasswordCheck(e)} marginBottom='5px' type="password" placeholder="password confirmation" width={'60%'}/>
-                {passwordError && <FormControl.ErrorMessage>Passwords do not match</FormControl.ErrorMessage>}
-                {signUpError && <FormControl.ErrorMessage>Signup error</FormControl.ErrorMessage>}
-                {sendEmailError && <FormControl.ErrorMessage>Could not send email verification.</FormControl.ErrorMessage> }
+                <Input isRequired onChangeText={(e) => setPasswordCheck(e)} 
+                    marginBottom='5px' type="password" placeholder="password confirmation" width={'60%'}/>
+                {passwordError && 
+                    <FormControl.ErrorMessage>Passwords do not match</FormControl.ErrorMessage>}
+                {signUpError && 
+                    <FormControl.ErrorMessage>Signup error</FormControl.ErrorMessage>}
+                {sendEmailError && 
+                    <FormControl.ErrorMessage>Could not send email verification.</FormControl.ErrorMessage> }
                 <HStack space={'1rem'}>
                     <Button onPress={() => navigate('/')}><Text variant={'button'}>Back</Text></Button>
                     <Button onPress={attemptSignup} ><Text variant={'button'}>Sign Up</Text></Button>
@@ -103,7 +111,7 @@ const Signup =  () => {
             </FormControl>
         </Box>
         
-    )
-}
+    );
+};
 
 export default Signup;
